Mermaid.jl

Summary / README

Mermaid.Mermaid — Module

Mermaid.jl

Mermaid.jl is a general purpose component-based simulation tool in Julia. It allows users to connect arbitrary Julia models from a wide range of packages to produce complex coupled simulations.

Any simulation that can be performed in Julia can be included as a component, from DifferentialEquations.jl, to Agents.jl and beyond.

Simple interface for extending components to tools not yet supported.

Installation (Coming soon)

Mermaid can be installed from Julia with:

using Pkg; Pkg.add("Mermaid")

FAQ

Q: How can I use Mermaid to run a hybrid model simulation?

A: Check out the Tutorial, it covers how to set up and run a hybrid simualtion between an Agent-based model and an ODE system.

Q: What if Mermaid doesn't have a component for the type of model I want to use?

A: Check out the Mermaid Interface, you can easily define new components for many different simulation tools.

Q: The Tutorial didn't cover what I wanted. Where can I view more examples?

A: In the Examples section of course. We have examples showing all of the more advanced features of Mermaid including:

duplicated components,

out of sync computation,

external (non-Julia) components,

surrogate model approximations,

mapping between spatial components on different resolutions,

integration with ModelingToolkit.jl.

Q: Is Mermaid the right tool for me? A: That depends on what type of hybrid simulation you are looking to run. Luckily, you can check out Is Mermaid right for me? and find out.