Mermaid Interface

In this section, we will:

See the integrator interface that Mermaid uses,

Find out how this fits with CommonSolve,

Look at how the AgentsComponent is defined, as an example.

Interface requirements

The interface used is Mermaid is compatible with, and uses, the SciML CommonSolve interface. In particular, each Component is some immutable problem type that stores required data to solve a Component (for example, an ODEProblem). We then also have an Integrator for each Component which stores the current state and handles solving over time.

Component

There are two categories of Components in Mermaid. AbstractTimeDependentComponents and AbstractTimeIndependentComponents, both of which are AbstractComponents.

A component must contain the following fields:

name which is a subtype of AbstractString,

time_step which is a Float64,

state_names which is a Dict which maps from variable names in the form of a String (variable name format) to some other object (typically a numerical index or symbol) which can be used to identify that variable internally.

Integrator

A ComponentIntegrator is a mutable struct which can be freely modified over a simulation, since it will be typically discarded at the end. It typically only stores the current state of that component, and has some associated functions for handling this state.

In addition to having some functions defined on the integrator, we also have some required fields.

outputs and inputs are Dicts for mapping from a ConnectedVariable to its current value. The values in these Dicts are set by Mermaid but the initial creation of this Dict (in particular, its keys) need to be defined in the init function. The values from inputs should be used to set the state during the step! function.

The original Component should also be stored under the field name component. This allows access to any problem settings, or other fields like the name.

setstate! and inputs

Couldn't the state be set by mermaid, rather than in the step function?

Warning

An Integrator is required for all Components, even TimeIndependentComponents. They just store the state that they were last called with.

Functions

step!

The step! function should advance the ComponentIntegrator (its only input) one time step (defined by Component.time_step). This should also include reading the inputs field and updating the integrators internal state accordingly.

init

The init function takes as inputs a Component and a Vector{Connector} and returns the corresponding ComponentIntegrator. This includes the process of creating the inputs and outputs Dicts which require the keys to be specified, i.e. the init function should read through the Vector{Connector} and look for any outputs or inputs into this component, and add the ConnectedVariables as keys to the inputs and outputs Dicts.

getstate and setstate!

The function getstate function reads the state of a ComponentIntegrator at a given ConnectedVariable and returns it. The setstate! function similarly mutates the current state of the ComponentIntegrator to assign an inputted value to the ConnectedVariable.

gettime

Simply returns the current simulated time of a ComponentIntegrator.

AgentsComponent Example

Code example

We should include the code of the Agents component here too. probably with literate

Mermaid.AgentsComponent — Type

AgentsComponent <: AbstractTimeDependentComponent

A component that represents an agent-based model (ABM) using the Agents.jl package.

Fields

model::StandardABM: The agent-based model to be solved.

name::String="Agents Component": The name of the component.

state_names::Dict{String,Any} = Dict{String,Any}(): A dictionary mapping ConnectedVariable names (as strings) to their corresponding properties in the agent model. The properties differentiate between :model properties and :agent properties.

time_step::Float64=1.0: The time step for the component (not the ABM solver timestep), i.e. how frequently should the inputs and outputs be updated.

Mermaid.AgentsComponentIntegrator — Type

AgentsComponentIntegrator <: ComponentIntegrator

A mutable struct that integrates an AgentsComponent using the Agents.jl package.

Fields

integrator::StandardABM: The agent-based model integrator.

component::AgentsComponent: The AgentsComponent being integrated.

outputs::Dict{ConnectedVariable,Any}: A dictionary mapping ConnectedVariable names to their initial values from the component.

inputs::Dict{ConnectedVariable,Any}: A dictionary mapping ConnectedVariable names to their current values (initially 0).

CommonSolve.step! — Method

step!(compInt::AgentsComponentIntegrator)

Sets the state based on the current inputs and steps the AgentsComponentIntegrator in time one step.

Arguments

compInt::AgentsComponentIntegrator: The component integrator to be stepped. Its internal state will be mutated.

CommonSolve.init — Method

init(c::AgentsComponent, conns::Vector{Connector})

Initializes an AgentsComponentIntegrator for the given AgentsComponent and its connections.

Arguments

c::AgentsComponent: The AgentsComponent to be integrated.

conns::Vector{Connector}: The Connectors that define the inputs and outputs of the component.

Returns

AgentsComponentIntegrator: The initialized integrator for the Agents.jl component.

Mermaid.getstate — Method

getstate(compInt::AgentsComponentIntegrator, key::ConnectedVariable)

Retrieves the state of a specific variable from the AgentsComponentIntegrator.

Arguments

compInt::AgentsComponentIntegrator: The component integrator containing the agent-based model.

key::ConnectedVariable: The ConnectedVariable specifying which variable's state to retrieve.

Returns

The current state of the variable specified by key, which can be a model-level property or an agent-specific property.

Mermaid.setstate! — Method

setstate!(compInt::AgentsComponentIntegrator, key::ConnectedVariable, value)

Sets the state of a specific variable in the AgentsComponentIntegrator.

Arguments

compInt::AgentsComponentIntegrator: The component integrator containing the agent-based model.

key::ConnectedVariable: The ConnectedVariable specifying which variable's state to set.

value: The value to assign to the specified variable's state.

Mermaid.gettime — Method

gettime(compInt::AgentsComponentIntegrator)

Returns the simulation time of the AgentsComponentIntegrator at the current state.

Arguments

compInt::AgentsComponentIntegrator: The component integrator for which to retrieve the current time.

Returns

time: The current simulation time.

Variable Names