Tutorial

In this tutorial, we will:

Create a hybrid simulation between an Agent-based model, defined in Agents.jl, and an ODE system defined through DifferentialEquations.jl.

Introduce Mermaid Components for Agents.jl and DifferentialEquations.jl.

Demonstrate how these Components can be connected together through Connectors.

Solve the hybrid model.

Visualise the results of the simulation.

The example system we will use for this will be a model of a forest fire (governed by an Agent-based model), with the growth of each tree informed by an ODE model.

Components

To begin, we need to define our components. These will be an ODE model component for each tree, and an Agent-based model component for handling the forest level properties (in this case, the spread of heat/fire).

ODE Components

To define the ODE model, let's have a look at how to define an ODE Component.

Mermaid.DEComponent — Type

DEComponent(model::DiffEqBase.AbstractDEProblem, alg;
            name::String="DE Component", timestep::Float64=1.0, intkwargs::Tuple=(),
            state_names::Dict{String,Any}=Dict{String,Any}())
DEComponent(model::DiffEqBase.AbstractDEProblem; kwargs...)

Arguments

model::DiffEqBase.AbstractDEProblem: SciML DE problem (e.g., ODEProblem, etc.)

alg: Algorithm from DifferentialEquations.jl to be used for solving the DEProblem. If no algorithm is provided, the algorithm will be automatically chosen by DifferentialEquations.jl.

Keyword Arguments

name::AbstractString: Name of the component. Defaults to "DE Component".

timestep::Real: Time step for the component. Defaults to 1.0.

intkwargs: Additional keyword arguments for the DE solver. Defaults to no keywords.

state_names: Dictionary mapping variable names (as strings) to their corresponding indices in the state vector or symbols from Symbolics.jl. Defaults to an empty dictionary.

We see that we need to define an ODEProblem to use in the component, so let's create one.

using OrdinaryDiffEq
function tree!(du, u, p, t)
    heat, life = u
    du[1] = 0
    du[2] = (life*(1-life/10.0)-heat*life)/10
end
u0 = [4.0, 2.0]
tspan = (0.0, 150.0)
prob = ODEProblem(tree!, u0, tspan)

Next, we want to wrap this ODEProblem inside an DEComponent. For this, we will need to define the state_names field, and should generally provide a value for the name field (since component names in a hybrid simulation should be unique).

using Mermaid
comp1 = DEComponent(prob, Rodas5P();
    name="tree", state_names=Dict("heat" => 1, "life" => 2),
)

Duplicated Components

Since the ODEProblem is defined for only a single tree, we can efficiently simulate time ODE system many times by generating a duplicated component. This component stores a single ODEProblem that it will solve across many different states. In this case, we can have a state for each tree in the Agent-based model. Let's have a look at how to define a DuplicatedComponent.

Mermaid.DuplicatedComponent — Type

DuplicatedComponent <: AbstractComponent

Represents a component that is duplicated in the simulation, allowing a single component to have multiple states.

Fields

component::AbstractTimeDependentComponent: The original component to be duplicated.

instances::Union{Int,Nothing}: Number of instances of the component. If nothing, then the number is variable and determined by the simulation.

name::String: Name of the duplicated component.

init_states::Vector: Vector of states for the duplicated component, where each state corresponds to a particular instance.

dup_comp = DuplicatedComponent(comp1, [copy(u0) for _ in 1:640];
    instances=640,
)

Agents.jl Components

Now that we have created our DEComponent, we can move on to the AgentsComponent, so let's have a look at its documentation.

Mermaid.AgentsComponent — Type

AgentsComponent(model::StandardABM; name="Agents Component",
                state_names=Dict{String,Any}(), timestep::Real=1.0)

A Mermaid component that wraps an agent-based model (ABM) using the Agents.jl package.

Arguments

model::StandardABM: The agent-based model to be solved.

Keyword Arguments

name::AbstractString: The name of the component. Defaults to "Agents Component".

state_names: A dictionary mapping variable names (as strings) to their corresponding properties (agent properties or model properties) in the model. Defaults to an empty dictionary.

timestep::Real: The time step for the component (not the ABM solver timestep), i.e. how frequently should the inputs and outputs be updated.

We can see that, again, we need to define the model (this time a StandardABM from Agents.jl), a name and a state_names.

using Agents, Random, Statistics
@agent struct Tree(GridAgent{2})
    heat::Float64 # Heat is averaged across neighbors, passed to ODE model
    life::Float64 # Life is informed by ODE model
end

function forest_fire(; density=0.4, griddims=(40, 40), seed=2)
    space = GridSpaceSingle(griddims; periodic=false, metric=:chebyshev)
    rng = Random.MersenneTwister(seed)
    forest = StandardABM(Tree, space; rng, agent_step! = tree_step!)
    for _ in 1:floor(density * prod(griddims))
        # Randomly place trees in the grid
        add_agent_single!(forest; heat=rand(), life=rand())
    end
    return forest
end

function tree_step!(tree, forest)
    nearbyheat = mean([getproperty(neighbor, :heat) for neighbor in nearby_agents(tree, forest, 1)])
    if isnan(nearbyheat)
        nearbyheat = 0.0
    end
    tree.heat = tree.heat * 0.9 + nearbyheat * 0.1
    if rand(abmrng(forest)) < 1e-4 # Random chance of fire
        tree.heat = 10.0
    end
    # Simulate tree life cycle
    if tree.heat > 1.0 && tree.life > 1.0
        # Tree on fire
        tree.heat += 1.0
    else
        # Tree not on fire so heat disappates
        tree.heat -= 0.05
    end
    if tree.heat < 0.0
        tree.heat = 0.0
    end
end

forest = forest_fire()

comp2 = AgentsComponent(forest;
    name="forest", state_names=Dict("heat" => :heat, "life" => :life)
)

AgentsComponent{StandardABM{GridSpaceSingle{2, false}, Main.Tree, Dict{Int64, Main.Tree}, Tuple{DataType}, typeof(Main.tree_step!), typeof(dummystep), typeof(Agents.Schedulers.fastest), Nothing, Random.MersenneTwister}, String, Dict{String, Symbol}, Int64}(StandardABM with 640 agents of type Tree
 agents container: Dict
 space: GridSpaceSingle with size (40, 40), metric=chebyshev, periodic=false
 scheduler: fastest, "forest", Dict("heat" => :heat, "life" => :life), 1)

Connections

We can now set up the connections between the variables in the two components.

Mermaid.Connector — Type

Connector <: AbstractConnector

Represents a connection between multiple ConnectedVariables, possibly with a transformation function.

Fields

inputs::Vector{<:AbstractConnectedVariable}: Input variables for the connector.

outputs::Vector{<:AbstractConnectedVariable}: Output variables for the connector.

func::Union{Nothing,Function}: Optional function to transform inputs to outputs.

The format for specifying a ConnectedVariable is given in Mermaid Interface, but in short, it is a string containing a component name and a variable/state name. It can also contain optional indices for only accessing part of a variable (given at the end of the ConnectedVariable), or for only accessing some subcomponents, such as for DuplicatedComponents (given at the end of the component name).

conn1 = Connector(inputs=["forest.heat[1:640]"], outputs=["tree[1:640].heat"])
conn2 = Connector(inputs=["tree[1:640].life"], outputs=["forest.life[1:640]"])

Connector(ConnectedVariable[ConnectedVariable{SubString{String}, SubString{String}, Nothing, UnitRange{Int64}}("tree", "life", nothing, 1:640)], ConnectedVariable[ConnectedVariable{SubString{String}, SubString{String}, UnitRange{Int64}, Nothing}("forest", "life", 1:640, nothing)], nothing)

Solving the hybrid model

To create the hybrid model, we need to create a MermaidProblem. We can then solve this using the CommonSolve interface.

Mermaid.MermaidProblem — Type

MermaidProblem <: AbstractMermaidProblem
MermaidProblem(;
    components::Vector{AbstractComponent},
    connectors::Vector{AbstractConnector},
    tspan::Tuple{Float64, Float64},
    timescales::Vector{Float64}=ones(length(components)))

Defines a Mermaid hybrid simulation problem.

Keyword Arguments

components::Vector{<:AbstractComponent}: Vector of Components.

connectors::Vector{<:AbstractConnector}: Vector of Connector.

tspan::Tuple{Float64, Float64}: The time span of the simulation.

timescales::Vector{Float64}=ones(length(components)): Timescales for each component. Each component's timescale will be multiplied by the component's time to convert it to the universal simulation time.

mp = MermaidProblem(components=[dup_comp, comp2], connectors=[conn1, conn2], tspan=tspan)
alg = MinimumTimeStepper()
sol = solve(mp, alg)

MermaidSolution{Vector{Any}, Dict{Any, Any}}(Any[0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0  …  141.0, 142.0, 143.0, 144.0, 145.0, 146.0, 147.0, 148.0, 149.0, 150.0], Dict{Any, Any}(ConnectedVariable{SubString{String}, SubString{String}, Nothing, Nothing}("tree", "life", nothing, nothing) => Any[Any[2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0  …  2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0], Any[2.0796100898653918, 1.9877328094357418, 2.027589835038657, 2.1415004685015933, 2.103663756282345, 2.0524236560433593, 2.110134647524075, 2.075694095329474, 2.0109840284823437, 2.0449502586897634  …  1.9679369562903242, 2.109295634235028, 1.9730365549755031, 1.9759715863437686, 2.1440028019681128, 2.034793174332959, 1.9704972089776045, 1.9886895075661, 2.1042394752787534, 2.1413418772486583], Any[2.172636339048418, 1.994818866584963, 2.0693081781689213, 2.2848153964403872, 2.220365720143091, 2.118775657411349, 2.22960904006859, 2.169671718376212, 2.031757528488746, 2.1027347825244758  …  1.954946067457683, 2.2224002190427665, 1.9688325342971367, 1.9707645452726514, 2.296694878505803, 2.0832452093039957, 1.953588868536111, 1.9947951593142146, 2.2164933486791414, 2.29677693005439], Any[2.2800224710970602, 2.0199828097672263, 2.1252861254121274, 2.4324326277541783, 2.35145904531974, 2.199482685207175, 2.3596229244427978, 2.2821134958205773, 2.0626152875258064, 2.173815793335627  …  1.9593012130584717, 2.3409483253487315, 1.985297966133814, 1.9831678851702224, 2.4598155397755286, 2.145363845378352, 1.9488136554793682, 2.0172843762699606, 2.33812653476475, 2.4680652919429362], Any[2.4029047447709257, 2.062565865817354, 2.1959397002196384, 2.5868329947926307, 2.4984391795629985, 2.2951886946224804, 2.501481757977836, 2.4135770466389475, 2.1039615435052004, 2.2588789345387448  …  1.97990576270748, 2.4665982895707645, 2.021205028444466, 2.012508756622512, 2.6351906387196222, 2.2214487106958503, 1.9559334306162992, 2.0556307915526104, 2.4705677448450194, 2.657087482338974], Any[2.542610120701696, 2.1224139169390854, 2.2819490960722204, 2.750471010305281, 2.6629471413768435, 2.4067395166508425, 2.65660035415433, 2.5649364227980604, 2.1563184291300836, 2.3588249825575596  …  2.0161677011644428, 2.601035415640154, 2.0760037021625974, 2.0585588534683863, 2.8247303704142293, 2.312059456241068, 1.9749164205515575, 2.109729442529334, 2.61532002856952, 2.8567056481982998], Any[2.7005797817378205, 2.199712622115865, 2.3842463324847776, 2.925775003656582, 2.846770725477644, 2.5351705806015534, 2.826505149723929, 2.737336683733609, 2.2203290592274234, 2.4747655805725266  …  2.0679140472456767, 2.7459797151402072, 2.1496975292624096, 2.1214742642909616, 3.0304272118601157, 2.417994740967549, 2.005927005621539, 2.179838702382216, 2.773970193811656, 3.0650605993713986], Any[2.878150052723893, 2.294966135307152, 2.5037999088131264, 3.115157182253837, 3.0514390545594363, 2.681695659412928, 3.0128344140468157, 2.932156587785751, 2.2967658593205487, 2.608017267167816  …  2.1353326279377387, 2.9031957331992517, 2.2426196231635993, 2.2017536727581835, 3.245695546609738, 2.540278493552058, 2.049320843069552, 2.266538685227621, 2.9481961242775983, 3.281630748911798], Any[3.076744094691859, 2.40894985026637, 2.6417510329936387, 3.3210266471199037, 3.2675002177315604, 2.847525181983919, 3.217081989610894, 3.1435928250665186, 2.386539152358614, 2.760088955572956  …  2.218933090554834, 3.074502403647999, 2.355469945409634, 2.3002091905445985, 3.468645158386063, 2.6801500004663596, 2.105644361856066, 2.3706067009862086, 3.1397709368373015, 3.505767567047574], Any[3.293747688438022, 2.5425879634474082, 2.7993999005257675, 3.545800284485399, 3.4911734358584376, 3.03393015264253, 3.4390668735219894, 3.363021001444579, 2.4907054322295967, 2.9325051554557073  …  2.3195168356670672, 3.261781140146176, 2.4889987393271182, 2.417931881741127, 3.6985233986330486, 2.8390560244701386, 2.1756373828879987, 2.492972330410312, 3.3505627991839715, 3.7366984943523134]  …  Any[6.781427613374383, 9.999943709172339, 9.999951515356598, 9.999966314771433, 3.412281599644541, 1.6760201350021331e-9, 1.59093411601759e-5, 9.99996347820535, 9.999937274209385, 0.902860809071779  …  1.755031533502252e-9, 9.999961382920633, 9.999942313943631, 9.999938606592412, 9.999968469894787, 2.6172020266639216e-11, 9.999918950700817, 9.999941080802559, 9.999963273599667, 9.999968981074744], Any[6.899289246102027, 9.999949065925554, 9.999956129260209, 9.999969520334988, 2.2019608129251202, 1.2296239822256764e-9, 1.3253229861220764e-5, 9.999966953702144, 9.999943243323697, 0.9884299023699075  …  1.1720616153213769e-9, 9.999965057808772, 9.999947803469047, 9.999944448915137, 9.999971470372449, 2.529081321427681e-11, 9.999926663504832, 9.999946687675624, 9.999966768567132, 9.999971932907476], Any[7.023372419530936, 9.999953912921246, 9.999960304096508, 9.999972420850606, 1.3199323547604342, 9.069638276833942e-10, 1.1112969990814719e-5, 9.99997009846377, 9.999948644407823, 1.0811450383860712  …  7.888002691154769e-10, 9.999968382987396, 9.999952770602238, 9.999949735273232, 9.999974185318461, 2.4582310935361826e-11, 9.999933642348756, 9.99995176098959, 9.999969930946575, 9.999974603837686], Any[7.151713525655348, 9.999958298668364, 9.999964081647608, 9.999975045347123, 0.73712692413093, 6.725615551710394e-10, 9.377826401132111e-6, 9.999972943963463, 9.999953531515864, 1.1814167529922586  …  5.349427253653896e-10, 9.999971391735343, 9.999957265054466, 9.999954518572657, 9.999976641904471, 2.4032182028485447e-11, 9.999939957076268, 9.999956351518337, 9.99997279238755, 9.999977020596509], Any[7.282240625532893, 9.999962267059782, 9.999967499719654, 9.999977420090962, 0.384565422749472, 5.014163243404091e-10, 7.962840913274146e-6, 9.999975518679456, 9.999957953558201, 1.2896435277799982  …  3.6554819106721007e-10, 9.999974114164626, 9.999961331806496, 9.999958846684905, 9.999978864716454, 2.3629295670393148e-11, 9.999945670884875, 9.999960505204147, 9.999975381527827, 9.997035803442458], Any[7.412827479923287, 9.999965857811539, 9.99997059252115, 9.999979568849014, 0.2092973266837429, 3.758280860984801e-10, 6.802409327767586e-6, 9.99997784837997, 9.99996195479094, 1.4062037582738927  …  2.516827773605898e-10, 9.999976577521789, 9.999965011558754, 9.999962762926042, 9.99998087600076, 2.3365345478745162e-11, 9.99995084095833, 9.99996426361746, 9.999977724279983, 9.965131069684741], Any[7.54188338609936, 9.99996910686031, 9.99997339100532, 9.999981513126501, 0.1182306744317107, 2.832072392127166e-10, 5.845556726759453e-6, 9.999979956381102, 9.999965575258804, 1.5314466664024642  …  1.7458587503345883e-10, 9.99997880646057, 9.999968341138619, 9.999966306490206, 9.999982695886755, 2.3234567173387916e-11, 9.999955519038853, 9.999967664372898, 9.999979844090742, 9.885381703845464], Any[7.6522670358505005, 9.999972046723029, 9.999975923179857, 9.999983272382172, 0.06867801066634377, 2.1455759242873507e-10, 5.052437724715519e-6, 9.999981863780167, 9.999968851195858, 1.6656822309909054  …  1.2200777526879354e-10, 9.999980823288634, 9.99997135386898, 9.999969512841815, 9.999984342588272, 2.323353667555859e-11, 9.999959751944926, 9.999970741505658, 9.99998176217561, 9.745325744374652], Any[7.730299809236531, 9.99997470682231, 9.999978214387236, 9.999984864223066, 0.04070443272505285, 1.634210309002128e-10, 4.391726673381596e-6, 9.999983589666842, 9.999971815388129, 1.8091702739291222  …  8.589453364750102e-11, 9.999982648190837, 9.999974079901707, 9.999972414070502, 9.999985832585887, 2.336103896232349e-11, 9.999963582039817, 9.999973525812154, 9.999983497731202, 9.538278227936665], Any[7.765725365505342, 9.999977113780897, 9.99998028755831, 9.999986304580709, 0.024451880608020404, 1.2514014620876944e-10, 3.83865895093936e-6, 9.999985151314197, 9.999974497501723, 1.9621089132414937  …  6.091450183500175e-11, 9.999984299431205, 9.9999765465194, 9.999975039212227, 9.999987180791866, 2.3618001083571764e-11, 9.999967047655517, 9.999976045158188, 9.999985068127366, 9.26260047020216]], ConnectedVariable{SubString{String}, SubString{String}, Nothing, Nothing}("tree", "heat", nothing, nothing) => Any[Any[4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0  …  4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0, 4.0], Any[0.4056994708920292, 0.8621408571954849, 0.6616126907308237, 0.1093856021447891, 0.2895098423219379, 0.538639413965653, 0.25847781536337067, 0.4247416640213063, 0.7446795829873095, 0.5754903842255096  …  0.9632269394950569, 0.26249607979379186, 0.937089946306283, 0.9220774286060864, 0.09758473280477331, 0.625789503266974, 0.9500964669592646, 0.857280915546428, 0.286745065604205, 0.1101339712456395], Any[0.3497959850313117, 0.7652859002895189, 0.5914898216858787, 0.13095018955359727, 0.24391550126936445, 0.47327573810032236, 0.23230112228923816, 0.34494703650025177, 0.6950909281393338, 0.5139660891267129  …  0.8700902655022237, 0.2611068012826764, 0.8242374065786221, 0.8290507475262149, 0.09005910111709965, 0.5587723173616914, 0.8899774230237123, 0.7701700578056833, 0.2642690979982029, 0.07741480693871126], Any[0.2949506523095099, 0.6739008470713792, 0.5233519268064059, 0.1381195371239467, 0.19780560931977975, 0.4102598258584325, 0.2038191785582441, 0.27217611444498596, 0.6445445851976435, 0.4537267126372872  …  0.7820342099987054, 0.2521781426043584, 0.7190092363212658, 0.7395625169408644, 0.07608258884219153, 0.49475093015283994, 0.829353955661512, 0.6872856574995521, 0.2380639922312799, 0.042548993092919], Any[0.24095324457196782, 0.5872550025873265, 0.4569090049485745, 0.13365853724822307, 0.15124762030902827, 0.34935377377589183, 0.17317073919849613, 0.20517197645867852, 0.5931979962524601, 0.3945308611830369  …  0.6984244329770226, 0.23681348369712107, 0.6204255143407574, 0.6533491415097359, 0.056621461573636664, 0.4331636996860057, 0.7682943839369256, 0.6080485852097542, 0.20861958137507103, 0.0058579326442458335], Any[0.18763282041993212, 0.5047217671004359, 0.3919182489579908, 0.11980377308195093, 0.10431023731127097, 0.2903471134255444, 0.1405004923742531, 0.14288171691452384, 0.5411842356536942, 0.3361820065450534  …  0.6187029447840557, 0.21595326476414062, 0.5276337610530785, 0.5702067814956655, 0.03249665788741668, 0.3735465496137675, 0.7068705147607216, 0.531964734118281, 0.1763587055726823, 0.0], Any[0.135132101819687, 0.4261763504747926, 0.32817645582287563, 0.09836419526702812, 0.057056003176950795, 0.2330537795383753, 0.1059559230285423, 0.08442428857148694, 0.4886408245543363, 0.2785227776653319  …  0.5423794278047208, 0.19039927501110715, 0.43989938946040286, 0.48995214629100664, 0.004405549107862103, 0.3155168573615855, 0.6451544902342429, 0.45861644820027997, 0.14164667446012968, 0.0], Any[0.08430385509296144, 0.35136825264791827, 0.26635738176824686, 0.07080264028072249, 0.010877351631743228, 0.1773096645095314, 0.06968399085984922, 0.02906338775391265, 0.43568667243190107, 0.22144499103638932  …  0.4690234871817515, 0.1608352933276081, 0.3572248151672399, 0.41241986942887915, 0.0, 0.2587599404787417, 0.5832166012814555, 0.387663496080575, 0.10479935460542915, 0.0], Any[0.03507463487312722, 0.28009972115536896, 0.2064302729874381, 0.0383018334812186, 0.0, 0.12357410337470535, 0.03262847769110007, 0.0, 0.38242313744034334, 0.16490604852842905  …  0.39825771834471513, 0.1278446597281836, 0.2791640461073816, 0.3374602365778441, 0.0, 0.20302017479847173, 0.5211238323404401, 0.31924100705773695, 0.0660900869180719, 0.0], Any[0.0, 0.2125402679364557, 0.14835180494922295, 0.0018178274390837357, 0.0, 0.07189405771146411, 0.0, 0.0, 0.32893655354722545, 0.10947496013394252  …  0.32977151298631374, 0.09192526232584124, 0.20640839832132946, 0.26499367641526295, 0.0, 0.14809095062923655, 0.45894109249300324, 0.2535496615043272, 0.02575560773986564, 0.0]  …  Any[0.16400026217519342, 0.0, 0.0, 0.0, 4.194240970266494, 4.150664342307108, 2.8938872228996724, 0.0, 0.0, 0.0  …  5.115042578169678, 0.0, 0.0, 0.0, 0.0, 1.4014063482425354, 0.0, 0.0, 0.0, 0.0], Any[0.14360633338725404, 0.0, 0.0, 0.0, 5.104301627316801, 4.097135988360083, 2.8266499987795415, 0.0, 0.0, 0.0  …  5.037225725859947, 0.0, 0.0, 0.0, 0.0, 1.342496956389519, 0.0, 0.0, 0.0, 0.0], Any[0.12556230775087662, 0.0, 0.0, 0.0, 5.94546266401363, 4.043611283004962, 2.761282705314569, 0.0, 0.0, 0.0  …  4.960064064507046, 0.0, 0.0, 0.0, 0.0, 1.284140991142033, 0.0, 0.0, 0.0, 0.0], Any[0.11010449171718024, 0.0, 0.0, 0.0, 6.725777630214801, 3.990089291056859, 2.6976478244995175, 0.0, 0.0, 0.0  …  4.883534669317904, 0.0, 0.0, 0.0, 0.0, 1.2263326664266199, 0.0, 0.0, 0.0, 0.0], Any[0.09737735309997202, 0.0, 0.0, 0.0, 7.452304422279733, 3.9365689660378584, 2.6356208788684796, 0.0, 0.0, 0.0  …  4.807615728256618, 0.0, 0.0, 0.0, 0.0, 1.1690656580434708, 0.0, 0.0, 0.0, 0.003093472778653239], Any[0.08745248175571772, 0.0, 0.0, 0.0, 7.0547807621758, 3.8830492197871855, 2.575088988466701, 0.0, 0.0, 0.0  …  4.7322865204201765, 0.0, 0.0, 0.0, 0.0, 1.1123332196114415, 0.0, 0.0, 0.0, 0.03388426608461419], Any[0.07960266289777725, 0.0, 0.0, 0.0, 6.695241459367954, 3.8295289835546957, 2.5159496046705563, 0.0, 0.0, 0.0  …  4.657527389945749, 0.0, 0.0, 0.0, 0.0, 1.0561282806097876, 0.0, 0.0, 0.0, 0.08789600832544166], Any[0.0949358457124705, 0.0, 0.0, 0.0, 6.422964864931978, 3.7760072609118165, 2.458109398007511, 0.0, 0.0, 0.0  …  4.583319716005155, 0.0, 0.0, 0.0, 0.0, 1.0004435291043705, 0.0, 0.0, 0.0, 0.16128891374174414], Any[0.12937125756638285, 0.0, 0.0, 0.0, 6.225574088073212, 3.7224831707472203, 2.4014832801328385, 0.0, 0.0, 0.0  …  4.509645879503848, 0.0, 0.0, 0.0, 0.0, 0.9452714813810397, 0.0, 0.0, 0.0, 0.2507708356405692], Any[0.17945501962980187, 0.0, 0.0, 0.0, 6.093110342642303, 3.6689560399117305, 2.3459935427314313, 0.0, 0.0, 0.0  …  4.436489227118252, 0.0, 0.0, 0.0, 0.0, 0.8906045403946548, 0.0, 0.0, 0.0, 0.35352080489586735]], ConnectedVariable{SubString{String}, SubString{String}, Nothing, Nothing}("forest", "heat", nothing, nothing) => Any[[0.31697297370953315, 0.9019442852469484, 0.552292202884825, 0.010586336990597855, 0.4221532149351479, 0.857280915546428, 0.018549043254140973, 0.9032025955054463, 0.5302247341105649, 0.36120456779231924  …  0.023836707037078853, 0.9566432215786556, 0.6852846949618958, 0.48140335105572807, 0.31585364917166225, 0.8297800254674621, 0.2929970720203482, 0.6177724286804258, 0.7331476100918907, 0.6365344676359204], [0.25106590083937763, 0.8061662483084966, 0.5018642515818463, 0.0, 0.36445953381616225, 0.7701700578056833, 0.0, 0.7923975536001365, 0.42720226069950845, 0.34286122438641015  …  0.0005976427719493055, 0.8350939390064527, 0.6063531362093015, 0.3981546567617196, 0.289576938508048, 0.7369821390684467, 0.2715981021615615, 0.5730481806661835, 0.6669785586611509, 0.5871482917366181], [0.18869872422353418, 0.7165746095407494, 0.45194851857513213, 0.0, 0.309725312529085, 0.6872856574995521, 0.0, 0.6893760678111059, 0.3344820346295576, 0.3195925382998657  …  0.0, 0.7249512432135046, 0.5306829211135023, 0.31933024012727473, 0.2603800949821761, 0.6487964346332189, 0.24502057667431215, 0.5260121811784484, 0.6016916141687156, 0.5360579309162993], [0.1295631235490614, 0.6322317491051355, 0.4023319370346784, 0.0, 0.2573034997041278, 0.6080485852097542, 0.0, 0.5930775641655952, 0.2510338311666019, 0.29208012100715436  …  0.0, 0.6243862627661987, 0.457823107116165, 0.2443754582537564, 0.22858310863079434, 0.564735545914177, 0.21406037848422094, 0.4770884225399091, 0.5372083905412219, 0.48358155576486855], [0.07353149151203547, 0.5523476371651597, 0.35285715741518314, 0.0, 0.2066678679141975, 0.531964734118281, 0.0, 0.5025884650743144, 0.17593044804994168, 0.2609115050705132  …  0.0, 0.531860738592217, 0.38738827854286423, 0.17281836238840564, 0.19447223166625022, 0.48437503568529733, 0.17941340237129177, 0.42663994495754604, 0.47346013172344364, 0.4299881933113681], [0.020236300864191523, 0.47625609412225794, 0.303408450792269, 0.0, 0.15739151840763121, 0.45861644820027997, 0.0, 0.41744718580032636, 0.10833740324494752, 0.2265989629888267  …  0.0, 0.44608076549026454, 0.3190489080118634, 0.10602818924582484, 0.15830326717164261, 0.40734450273448214, 0.14167960405230107, 0.37497865027474203, 0.4103947859815889, 0.3755054782601294], [0.0, 0.4033949149885359, 0.25390243750889835, 0.0, 0.10912914833718114, 0.387663496080575, 0.0, 0.3371936498533347, 0.04750366292045277, 0.18959089917780608  …  0.0, 0.3659594730165403, 0.252524021308073, 0.04542537032124236, 0.12030458964663605, 0.3347883930211716, 0.10137452799195941, 0.32237891779562283, 0.3480143964076904, 0.3203291495546873], [0.0, 0.33328922471181965, 0.20428155895034283, 0.0, 0.06160242502032415, 0.31924100705773695, 0.0, 0.261327750189273, 0.0, 0.15027614930721156  …  0.0, 0.290585262504523, 0.1880949114583061, 0.0, 0.08067990832760012, 0.2663298388857098, 0.05894125907209728, 0.2690739972212956, 0.28632311147685263, 0.2648890822886945], [0.0, 0.26553753544756653, 0.15450902006796663, 0.0, 0.014587925905276009, 0.2535496615043272, 0.0, 0.19096618505217466, 0.0, 0.10900434856232431  …  0.0, 0.21919389652936222, 0.12684332172018647, 0.0, 0.03961078648285625, 0.20161259672276843, 0.014761336791008009, 0.21527426937060473, 0.2253263775851556, 0.20988665395275324], [0.0, 0.19980006439390918, 0.10456447267630041, 0.0, 0.0, 0.1904410015184988, 0.0, 0.12646351145412393, 0.0, 0.06608903931759726  …  0.0, 0.15116323901980017, 0.068623318899782, 0.0, 0.0, 0.14031374283348774, 0.0, 0.16116965811601625, 0.165048260106526, 0.15548181773017367]  …  [7.845237343488007, 0.0, 0.0, 4.1778510247134575, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 3.4038568262688127, 0.0, 2.4070061530291453, 0.0, 0.0, 0.0, 2.4308068537774083], [7.6632589955661485, 0.0, 0.0, 4.10980296874686, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 3.310941807520402, 0.0, 2.339420687176112, 0.0, 0.0, 0.0, 2.3623586098335068], [7.54680300758811, 0.0, 0.0, 4.04211559139785, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 3.2194451262257986, 0.0, 2.2720716505034724, 0.0, 0.0, 0.0, 2.2941953923310696], [7.487976927017802, 0.0, 0.0, 3.974790452208919, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 3.1293380273960585, 0.0, 2.2049626585241695, 0.0, 0.0, 0.0, 2.226320993008937], [7.4456264476419065, 0.0, 0.0, 3.9078291888694396, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 3.040591460180753, 0.0, 2.138097062853711, 0.0, 0.0, 0.0, 2.1587387840328964], [7.382457763179176, 0.0, 0.0, 3.841233487583423, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 2.953176272545101, 0.0, 2.0714779566078483, 0.0, 0.0, 0.0, 2.091451738189744], [7.303349630063357, 0.0, 0.0, 3.775005061006129, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 2.8670633674273507, 0.0, 2.0051081818243173, 0.0, 0.0, 0.0, 2.0244624502516535], [7.214135861425428, 0.0, 0.0, 3.7091455829059226, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 2.782223827274727, 0.0, 1.9389903383589417, 0.0, 0.0, 0.0, 1.9577731589699365], [7.120405748049786, 0.0, 0.0, 3.643657958009828, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 2.6986290126612724, 0.0, 1.873126793811757, 0.0, 0.0, 0.0, 1.8913857692674347], [7.026578956407458, 0.0, 0.0, 3.578546962810849, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 2.6162506396994867, 0.0, 1.8075196941259548, 0.0, 0.0, 0.0, 1.8253018742903802]], ConnectedVariable{SubString{String}, SubString{String}, Nothing, Nothing}("forest", "life", nothing, nothing) => Any[[0.9981524689156519, 0.04012657305194167, 0.52561035445006, 0.0034119548949205214, 0.29106598429109853, 0.2266802283321938, 0.01977093022608567, 0.06324343290801016, 0.2826281490429592, 0.7629345621932843  …  0.49920284686502947, 0.4767388645441094, 0.9910606892436803, 0.9617136222541233, 0.9571806266623273, 0.2885882175063139, 0.5263090812568187, 0.16218117856891057, 0.6714506470124705, 0.04972326238325486], [2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0  …  2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0, 2.0], [2.0979535009631505, 1.9799145384004018, 2.049651704787127, 2.162540329886198, 2.076225975073137, 1.9886895075661, 2.160837048934521, 1.9796678772725755, 2.0541339584750826, 2.088789016320832  …  2.159706715132911, 1.9692202729484174, 2.0228437643843966, 2.0640849809742083, 2.098185934667032, 1.994111786770137, 2.102937821482343, 2.0364087280967857, 2.0132812713790043, 2.0326299085310233], [2.212901446922181, 1.979099611263906, 2.109995115566275, 2.3368287007132578, 2.1660303428013075, 1.9947951593142146, 2.335029043574682, 1.9815575555370455, 2.130194811442633, 2.1835208080954196  …  2.333696816460864, 1.9629871155392449, 2.061527285822239, 2.146460205542508, 2.204726275779982, 2.0067094033740567, 2.2135461782060766, 2.081921812915609, 2.039700413487215, 2.0752389627762087], [2.345843102424952, 1.9959282675891712, 2.1815459662830663, 2.520646102043129, 2.2699231905137216, 2.0172843762699606, 2.5187514112584877, 2.003746920277397, 2.2277175434963463, 2.285616226477153  …  2.5173487887637376, 1.9783543592157302, 2.115924413964936, 2.247724025751161, 2.320865492988511, 2.0368382994026675, 2.3334821507074768, 2.137407808001617, 2.079315298454028, 2.1285726807005116], [2.497960091320915, 2.0294384588359433, 2.2649869338069744, 2.7138022302133664, 2.388643307165932, 2.0556307915526104, 2.711814991836906, 2.045140925470973, 2.3466881057190356, 2.3965235728191123  …  2.710343807494463, 2.013485844746234, 2.186274910037724, 2.368840884276846, 2.4479257308497115, 2.0840659468338485, 2.4644515129423112, 2.2037918411094, 2.1324238816566883, 2.193467773623558], [2.670550485279313, 2.079196143100008, 2.3611623222619365, 2.915989158052377, 2.5231376997481023, 2.109729442529334, 2.9139130748125868, 2.1052960131319547, 2.4874185232575603, 2.51773387271445  …  2.9123760628436357, 2.0673702448130906, 2.2731451917724783, 2.5111057883076633, 2.5873161460861667, 2.1483958565423005, 2.6082208571834564, 2.282079082711319, 2.199552504064432, 2.2708722920106017], [2.8650615818253655, 2.1452032438693838, 2.471076590187723, 3.126775325118712, 2.674546272481951, 2.179838702382216, 3.1246153633829703, 2.184232979955927, 2.6504448980219877, 2.650794867885432  …  3.123016193667277, 2.139634644948807, 2.3774028784717305, 2.675648459664731, 2.7405412372588627, 2.2302084006347016, 2.7666310182910383, 2.3733749789570715, 2.2814481985442328, 2.3618637264329094], [3.073763732498698, 2.2278359975261046, 2.5958948967325894, 3.345602414095365, 2.8441889569600547, 2.266538685227621, 3.343364819176309, 2.2823694113988124, 2.8364378637023466, 2.797323609893876  …  3.34170810880923, 2.230420991607035, 2.5002004374399824, 2.863260327467216, 2.909207348015725, 2.3298816736972348, 2.9416082990544687, 2.4789012142697304, 2.37906586952155, 2.467664606550174], [3.290657032302164, 2.3278040286642443, 2.7369442069126038, 3.5717856135686765, 3.0335529762594904, 2.3706067009862086, 3.569477888847969, 2.400482469503228, 3.043944492246818, 2.959018060461522  …  3.5677691869254846, 2.3403032536733845, 2.6428286573916404, 3.0718877593178964, 3.0950263952384636, 2.4480634737091753, 3.1351713403380557, 2.6000110459522645, 2.493567956797348, 2.589589456997285]  …  [0.014379768553729008, 9.999932915746754, 9.999948614367053, 7.05102172126901e-8, 9.999957411326205, 9.99993488425682, 9.99996666097275, 9.999938184052095, 9.999957710036862, 9.999954679461476  …  9.999966636142853, 9.99993537615591, 9.99994696520204, 6.695901184128134e-8, 9.999958472332487, 1.2210542845179871e-17, 9.999959390452803, 9.999943202595036, 9.999939971986963, 5.857771638618743e-14], [0.007037412560566073, 9.99993929961875, 9.999953504333824, 5.096462650084843e-8, 9.999961464158746, 9.999941080802559, 9.999969833591093, 9.9999440665844, 9.999961734443541, 9.99995899226325  …  9.999969811124057, 9.99994152589181, 9.999952012106132, 5.2157022315339403e-8, 9.999962424197701, 1.0536204033462799e-17, 9.999963254947966, 9.999948607554964, 9.999945684376646, 5.0420647730772134e-14], [0.003549001082826103, 9.999945075992029, 9.999957928962854, 3.708997410623997e-8, 9.999965131316111, 9.999946687675624, 9.999972704296617, 9.999949389325707, 9.999965375880087, 9.99996289465088  …  9.99997268396759, 9.999947090409483, 9.999956578738185, 4.101234609870256e-8, 9.999965999995908, 9.153320542940883e-18, 9.999966751690366, 9.999953498169988, 9.999950853166203, 4.3698775582987734e-14], [0.001822692171590405, 9.999950302676464, 9.99996193253613, 2.717687419111584e-8, 9.999968449499628, 9.99995176098959, 9.999975301819813, 9.999954205546091, 9.999968670790416, 9.999966425679853  …  9.999975283425337, 9.999952125398623, 9.999960710801338, 3.255005195728886e-8, 9.99996923551413, 8.005866981746061e-18, 9.999969915675837, 9.999957923385578, 9.999955530085003, 3.813315913149753e-14], [0.0009470732596581394, 9.999955031980823, 9.999965555121804, 2.004851007259198e-8, 9.999971451918134, 9.999956351518337, 9.999977652157158, 9.999958563446508, 9.99997165215044, 9.999969620689138  …  9.999977635513142, 9.999956681249564, 9.999964449649635, 2.6071283307195027e-8, 9.999972163133888, 7.049576052127663e-18, 9.99997277857001, 9.9999619274896, 9.999959761939907, 3.350399491145358e-14], [0.0004950070471889461, 9.999959311236221, 9.999968832975135, 1.4889790793187732e-8, 9.999974168620298, 9.999960505204147, 9.999979778831284, 9.999962506641142, 9.999974349798077, 9.999972511654853  …  9.99997976377115, 9.999960803557544, 9.999967832701882, 2.1071062408563232e-8, 9.99997481215527, 6.24931098153593e-18, 9.999975369025192, 9.999965550555508, 9.999963591083658, 2.9637266816640384e-14], [0.00025982437099336393, 9.999963183269783, 9.999971798901328, 1.1132768524142803e-8, 9.999976626795341, 9.99996426361746, 9.999981703126387, 9.999966074593873, 9.999976790731855, 9.999975127510243  …  9.999981689499412, 9.999964533578984, 9.999970893816116, 1.7181642258061832e-8, 9.999977209090146, 5.577058620301469e-18, 9.999977712967127, 9.999968828843375, 9.999967055838729, 2.639458050326132e-14], [0.00013724373031172936, 9.999966686833215, 9.999974482583845, 8.37934317096772e-9, 9.999978851045139, 9.999967664372898, 9.99998344430124, 9.999969303013204, 9.999978999381103, 9.999977494435264  …  9.999983431971037, 9.999967908644349, 9.999973663628431, 1.4133160941156667e-8, 9.9999793779275, 5.010389781526366e-18, 9.999979833854443, 9.999971795162757, 9.99997019088083, 2.3665387350368566e-14], [7.30703439129405e-5, 9.999969856990624, 9.999976910881445, 6.348820360219391e-9, 9.999980863630437, 9.999970741505658, 9.999985019781917, 9.999972224209614, 9.999980997850422, 9.99997963611855  …  9.999985008625089, 9.999970962531744, 9.999976169859577, 1.1726104837776133e-8, 9.999981340373507, 4.5312729053996285e-18, 9.99998175291342, 9.999974479201045, 9.999973027585924, 2.1361009757584705e-14], [3.925224430749797e-5, 9.999972725469403, 9.999979108096992, 4.842128844983658e-9, 9.99998268469362, 9.999973525812154, 9.999986445336217, 9.999974867418898, 9.999982806140931, 9.999981573994518  …  9.999986435241098, 9.999973725804939, 9.99997843759238, 9.811892362391005e-9, 9.999983116068744, 4.125156096958208e-18, 9.999983489350425, 9.999976907820558, 9.999975594344225, 1.941003053569171e-14]]))

Plotting the solution

After running solve, we get sol, a MermaidSolution instance. This stores all variables given in state_names at each timepoint.

Mermaid.MermaidSolution — Type

MermaidSolution{X, Y} <: AbstractMermaidSolution

Stores the solution of a MermaidProblem over time.

Fields

t::X: Time points at which the solution is saved.

u::Y: Dictionary mapping variables to their solution arrays.

using Plots

plot(sol.t, sol["forest.life[1]"], color=:green, label="Life")
plot!(sol.t, sol["forest.heat[1]"], color=:red, label="Heat")

Advanced Visualisations

While we can plot the variables from the ODE component easily, the Agent-based model is a bit more challenging. But default, we only store the variables given in state_names in the solution. This can be changed by providing save_vars=["forest.#model"] to solve, in which case the full Agent-based model state will be visable in the solution at all timepoints.

#model and Special Variables

"#model" is a special variable for AgentsComponents. Special variables, denoted by starting with # are not saved by default but can be used with connectors, getstate, setstate!, or save_vars. To view the special variables of a component, you can call variables(component).

However, this can be wasteful if we know we only want an animation of the model (which can be generated during simulation). We will set up a Connector which takes an input of the model's current state, and instead of a transformation, we will use a function which adds the current state to a video.

using Makie
using CairoMakie

groupcolor(tree) = tree.heat > 1 ? :red : :green
groupmarker(a) = a.life > 1 ? :utriangle : :circle
fig, ax = abmplot(forest; agent_color=groupcolor, agent_marker=groupmarker, agent_size=10)
io = VideoStream(fig)
function plot_input(model)
    empty!(ax)
    abmplot!(ax, model; agent_color=groupcolor, agent_marker=groupmarker, agent_size=10)
    recordframe!(io)
end

conn3 = Connector(
    inputs=["forest.#model"],
    outputs=Vector{String}(),
    func=(model) -> plot_input(model)
)

mp = MermaidProblem(components=[dup_comp, comp2], connectors=[conn1, conn2, conn3], tspan=tspan)
sol = solve(mp, alg)

save("forest_fire.mp4", io)

An animation of the forest fire simulation