API

Mermaid

Base.fullname — Method

fullname(var::AbstractConnectedVariable)

Return the full name of a ConnectedVariable as a string.

Arguments

var::AbstractConnectedVariable: The connected variable to get the full name for.

Returns

String: The full name of the connected variable.

Base.getindex — Method

Base.getindex(sol::AbstractMermaidSolution, var::AbstractString)
Base.getindex(sol::AbstractMermaidSolution, var::AbstractConnectedVariable)
Base.getindex(sol::AbstractMermaidSolution, index::Int)

Get the solution for a variable var or at a time index index from a MermaidSolution.

Arguments

sol::AbstractMermaidSolution: The solution object.

var::Union{AbstractString, AbstractConnectedVariable}: The variable name.

index::Int: The time index of the variable.

Returns

If var is provided, returns the MermaidSolution for that variable.

If index is provided, returns the MermaidSolution at time sol.t[index].

CommonSolve.init — Function

init(prob::MermaidProblem, alg::MermaidSolver; save_vars=[])
init(comp::AbstractComponent)

Initialises an integrator (MermaidIntegrator or AbstractComponentIntegrator) for the given MermaidProblem/AbstractComponent.

Arguments

prob::MermaidProblem: The hybrid problem to be solved.

alg::AbstractMermaidSolver: The AbstractMermaidSolver algorithm to be used for solving the problem.

comp::AbstractComponent: The component to be initialised.

Keyword Arguments

save_vars=[]: Which variables to save in the solution. Defaults to all non-special variables.

Returns

MermaidIntegrator: The initialized integrator for the problem.

CommonSolve.solve! — Method

solve!(merInt::AbstractMermaidIntegrator)

Solves the problem using the MermaidIntegrator. This handles all the message passing and calls step! on the MermaidIntegrator.

Arguments

merInt::AbstractMermaidIntegrator: The integrator to be solved.

Returns

MermaidSolution: The solution of the problem.

CommonSolve.step! — Function

step!(int::AbstractMermaidIntegrator)
step!(int::AbstractComponentIntegrator)

Advance the state of the integrator int by one time step.

Arguments

int::Union{AbstractMermaidIntegrator, AbstractComponentIntegrator}: The integrator to advance.

Mermaid.getstate — Function

getstate(comp::AbstractComponentIntegrator)
getstate(comp::AbstractComponentIntegrator, key)

Retrieve the state of a component.

Arguments

comp::AbstractComponentIntegrator: The component whose state is to be retrieved.

key: The key specifying which part of the component's state to retrieve.

Mermaid.gettime — Method

gettime(merInt::AbstractComponentIntegrator)

Get the current time of the integrator.

Arguments

int::AbstractComponentIntegrator: The integrator whose time is to be retrieved.

Returns

The current time of the integrator.

Mermaid.name — Method

name(int::AbstractComponentIntegrator)
name(comp::AbstractComponent)

Get the name of the integrator or component.

Mermaid.runconnection! — Method

runconnection!(merInt::AbstractMermaidIntegrator, conn::AbstractConnector)

Extract all the input states from merInt, apply the connection function, and set the output states in merInt.

Arguments

merInt::AbstractMermaidIntegrator: The Mermaid integrator containing the components.

conn::AbstractConnector: The connector defining the connection.

Mermaid.runconnection — Method

runconnection(merInt::AbstractMermaidIntegrator, conn::AbstractConnector)

Extract all the input states from merInt, apply the connection function, and return the output.

Arguments

merInt::AbstractMermaidIntegrator: The Mermaid integrator containing the components.

conn::AbstractConnector: The connector defining the connection.

Mermaid.setstate! — Function

setstate!(comp::AbstractComponentIntegrator, state)
setstate!(comp::AbstractComponentIntegrator, key, value)

Set the state of a component.

Arguments

comp::AbstractComponentIntegrator: The component whose state is to be set.

state: The new state to set for the entire component.

key: The key specifying which part of the component's state to set.

value: The value to set for the specified part of the component's state.

Mermaid.settime! — Method

settime!(merInt::AbstractComponentIntegrator, t)

Set the current time of the integrator.

Arguments

int::AbstractComponentIntegrator: The integrator whose time is to be set.

t: The time to set.

Mermaid.time_step — Method

time_step(int::AbstractComponent)
time_step(comp::AbstractComponentIntegrator)

Get the proposed time step of the integrator or component. It can depend on the current state.

Mermaid.update_solution! — Method

update_solution!(sol::MermaidSolution, merInt::MermaidIntegrator)

Update the MermaidSolution sol with the current time and state from the MermaidIntegrator.

Arguments

sol::MermaidSolution: The MermaidSolution to be updated.

merInt::MermaidIntegrator: The integrator object providing the current time (currtime) and states to access via getstate.

Mermaid.variables — Function

variables(comp::AbstractComponentIntegrator)
variables(comp::AbstractComponent)

Retrieve the variable names of a component.

Arguments

comp::Union{AbstractComponent, AbstractComponentIntegrator}: The component (or component integrator) whose variable names are to be retrieved.

Returns

A collection of variable names (as strings) associated with the component. This includes all special variables such as #time and #model if applicable.

Mermaid.variables — Method

variables(integrator::AbstractComponent)
variables(integrator::AbstractComponentIntegrator)

Get the variables names of the component or integrator that can be accessed through getstate and setstate!. This includes special variables like #time.

Mermaid.AbstractComponent — Type

abstract type AbstractComponent

Mermaid.AbstractComponentIntegrator — Type

abstract type AbstractComponentIntegrator

Mermaid.AbstractConnectedVariable — Type

abstract type AbstractConnectedVariable

Mermaid.AbstractConnector — Type

abstract type AbstractConnector

Mermaid.AbstractMermaidIntegrator — Type

abstract type AbstractMermaidIntegrator

Mermaid.AbstractMermaidProblem — Type

abstract type AbstractMermaidProblem

Mermaid.AbstractMermaidSolution — Type

abstract type AbstractMermaidSolution

Mermaid.AbstractMermaidSolution — Method

(sol::MermaidSolution)(t::Real)

Interpolates the solution at a given time t using linear interpolation.

Arguments

sol::MermaidSolution: The solution object containing time points and state histories.

t::Real: The time at which to interpolate the solution.

Returns

MermaidSolution: A new MermaidSolution object containing only that time and the interpolated time and state.

Mermaid.AbstractMermaidSolver — Type

abstract type AbstractMermaidSolver

Mermaid.AbstractTimeDependentComponent — Type

abstract type AbstractTimeDependentComponent <: AbstractComponent

Mermaid.AbstractTimeIndependentComponent — Type

abstract type AbstractTimeIndependentComponent <: AbstractComponent

Mermaid.ConnectedVariable — Type

ConnectedVariable <: AbstractConnectedVariable

Points to a variable within a component.

Fields

component::String: Name of the component.

variable::String: Name of the variable.

variableindex::Union{Nothing,Vector{Int},Int}: Index or range for the variable, if applicable.

duplicatedindex::Union{Nothing,Vector{Int},Int}: Index for duplicated components, if applicable.

Mermaid.ConnectedVariable — Method

ConnectedVariable(name::AbstractString)

Construct a ConnectedVariable from a string name.

Arguments

name::AbstractString: The full variable name.

Examples

ConnectedVariable("comp.var") ConnectedVariable("comp.var[1:5]") where 1:5 is the variableindex ConnectedVariable("comp[2].var") where 2 is the duplicatedindex ConnectedVariable("comp[1:3].var[4]")

Mermaid.Connector — Type

Connector <: AbstractConnector

Represents a connection between multiple ConnectedVariables, possibly with a transformation function.

Fields

inputs::Vector{<:AbstractConnectedVariable}: Input variables for the connector.

outputs::Vector{<:AbstractConnectedVariable}: Output variables for the connector.

func::Union{Nothing,Function}: Optional function to transform inputs to outputs.

Mermaid.Connector — Method

Connector(inputs, outputs; func=nothing)

Construct a Connector from string names for inputs and outputs.

Arguments

inputs::Vector{<:AbstractString}: Names of input variables.

outputs::Vector{<:AbstractString}: Names of output variables.

Keyword Arguments

func: Function for mapping inputs to outputs. Defaults to nothing which passes a single input, to all outputs.

Mermaid.DuplicatedComponent — Type

DuplicatedComponent <: AbstractComponent

Represents a component that is duplicated in the simulation, allowing a single component to have multiple states.

Fields

component::AbstractTimeDependentComponent: The original component to be duplicated.

instances::Union{Int,Nothing}: Number of instances of the component. If nothing, then the number is variable and determined by the simulation.

name::String: Name of the duplicated component.

init_states::Vector: Vector of states for the duplicated component, where each state corresponds to a particular instance.

Mermaid.DuplicatedComponent — Method

DuplicatedComponent(args...; kwargs...)

Duplicate an existing component into a DuplicatedComponent.

Arguments

component::AbstractTimeDependentComponent: The component to be duplicated.

init_states::AbstractVector: A vector of initial states for the duplicated component.

Keyword Arguments

instances::Union{Int, Nothing}: Number of instances of the component. If nothing, then the number is variable and determined by the simulation. Defaults to nothing.

name::AbstractString: Name of the duplicated component. Defaults to the original component's name.

time_step::Real: Time step for the duplicated component. Defaults to the original component's time step.

state_names: A dictionary mapping variable names (as strings) to their corresponding variables in the original component. Defaults to the original component's state names.

default_state: The default state to use when a new instance is created. Defaults to a zero vector of the same length as the first initial state.

Mermaid.MermaidIntegrator — Type

MermaidIntegrator <: AbstractMermaidIntegrator
MermaidIntegrator(
    integrators::Vector{AbstractComponentIntegrator},
    connectors::Vector{AbstractConnector},
    maxt::Float64,
    currtime::Float64,
    alg::AbstractMermaidSolver,
    save_vars::Vector{String})

Defines the integrator for a Mermaid hybrid simulation.

Fields

integrators::Vector{<:AbstractComponentIntegrator}: Vector of component integrators.

connectors::Vector{<:AbstractConnector}: Vector of Connector between components.

maxt::Float64: Maximum simulation time.

currtime::Float64: Current simulation time.

alg::AbstractMermaidSolver: The Mermaid solver algorithm to be used.

save_vars::Vector{String}: Variables to be saved during the simulation.

Mermaid.MermaidProblem — Type

MermaidProblem <: AbstractMermaidProblem
MermaidProblem(;
    components::Vector{AbstractComponent},
    connectors::Vector{AbstractConnector},
    max_t::Float64=1.0)

Defines a Mermaid hybrid simulation problem.

Keyword Arguments

components::Vector{<:AbstractComponent}: Vector of Components.

connectors::Vector{<:AbstractConnector}: Vector of Connector.

max_t::Float64=1.0: Maximum simulation time. Defaults to 1.0.

Mermaid.MermaidSolution — Type

MermaidSolution{X, Y} <: AbstractMermaidSolution

Stores the solution of a MermaidProblem over time.

Fields

t::X: Time points at which the solution is saved.

u::Y: Dictionary mapping variables to their solution arrays.

Mermaid.MermaidSolution — Method

MermaidSolution(int::MermaidIntegrator) <: AbstractMermaidSolution

Create a MermaidSolution object initialized for the save_vars/variables in the given MermaidIntegrator.

Arguments

int::MermaidIntegrator: The integrator to extract solution structure from.

Returns

MermaidSolution: A new MermaidSolution object with empty time and state arrays for each variable to be saved.

Mermaid.MinimumTimeStepper — Type

MinimumTimeStepper()

A solver that steps the integrators in a MermaidIntegrator advancing to the next event.

Extensions

Mermaid.AgentsComponent — Method

AgentsComponent(model::StandardABM; name="Agents Component",
                state_names=Dict{String,Any}(), time_step::Real=1.0)

A Mermaid component that wraps an agent-based model (ABM) using the Agents.jl package.

Arguments

model::StandardABM: The agent-based model to be solved.

Keyword Arguments

name::AbstractString: The name of the component. Defaults to "Agents Component".

state_names: A dictionary mapping variable names (as strings) to their corresponding properties (agent properties or model properties) in the model. Defaults to an empty dictionary.

time_step::Real: The time step for the component (not the ABM solver timestep), i.e. how frequently should the inputs and outputs be updated.

Mermaid.DEComponent — Method

DEComponent(model::DiffEqBase.AbstractDEProblem, alg;
            name::String="DE Component", time_step::Float64=1.0, intkwargs::Tuple=(),
            state_names::Dict{String,Any}=Dict{String,Any}())
DEComponent(model::DiffEqBase.AbstractDEProblem; kwargs...)

Arguments

model::DiffEqBase.AbstractDEProblem: SciML DE problem (e.g., ODEProblem, etc.)

alg: Algorithm from DifferentialEquations.jl to be used for solving the DEProblem. If no algorithm is provided, the algorithm will be automatically chosen by DifferentialEquations.jl.

Keyword Arguments

name::AbstractString: Name of the component. Defaults to "DE Component".

time_step::Real: Time step for the component. Defaults to 1.0.

intkwargs: Additional keyword arguments for the DE solver. Defaults to no keywords.

state_names: Dictionary mapping variable names (as strings) to their corresponding indices in the state vector or symbols from Symbolics.jl. Defaults to an empty dictionary.

Mermaid.MOLComponent — Method

MOLComponent(model::DiffEqBase.AbstractDEProblem, alg::DiffEqBase.AbstractDEAlgorithm;
             name::String="MOL Component", time_step::Real=1.0, intkwargs::Tuple=(),
             state_names::Dict{String,Any}=Dict{String,Any}())

Arguments

model::DiffEqBase.AbstractDEProblem: SciML DE problem (e.g., ODEProblem, etc.)

alg::DiffEqBase.AbstractDEAlgorithm: Algorithm from DifferentialEquations.jl to be used for solving the DEProblem.

Keyword Arguments

name::AbstractString: Name of the component. Defaults to "MOL Component".

time_step::Real: Time step for the component. Defaults to 1.0.

intkwargs: Additional keyword arguments for the DE solver. Defaults to no keywords.

state_names: Dictionary mapping variable names (as strings) to their corresponding indices in the state vector or symbols from Symbolics.jl. Defaults to an empty dictionary.

Mermaid.SurrogateComponent — Method

SurrogateComponent(args...; kwargs...)

Represents a component that is replaced with a surrogate in the simulation, speeding up computation of a complex step! function.

Arguments

component::AbstractTimeDependentComponent: The original component to be replaced with a surrogate.

surrogate: The surrogate model or method to use for the component.

lower_bound: Lower bounds for each state variable for surrogate sampling.

upper_bound: Upper bounds for each state variable for surrogate sampling.

Keyword Arguments

name::AbstractString: Name of the component. Defaults to the same as the original component.

time_step::Real: Time step for the component. Defaults to the same as the original component.

model: A Flux.jl model to use as the surrogate. If nothing, a default feedforward neural network is created.

state_names: Dictionary mapping variable names (as strings) to their corresponding indices in the state vector or symbols from ModelingToolkit/Symbolics. Defaults to the same as the original component.

n_samples::Integer: Number of samples to use for training the surrogate. Defaults to 1000.

n_epochs::Integer: Number of training epochs for the surrogate. Defaults to 1000.